+++ packages/canon/src/components/Grid/Grid.tsx
@@ -14,56 +14,50 @@
  * limitations under the License.
  */
 
-import { createElement } from 'react';
+import { createElement, forwardRef } from 'react';
 import { GridItemProps, GridProps } from './types';
-import { gridItemSprinkles, gridSprinkles } from './sprinkles.css';
+import { getClassNames } from '../../utils/getClassNames';
 
-/** @public */
-export const Grid = (props: GridProps) => {
+const GridBase = forwardRef<HTMLDivElement, GridProps>((props, ref) => {
   const {
     children,
-    columns,
     gap = 'xs',
+    columns = 'auto',
     className,
     style,
     ...restProps
   } = props;
 
-  const sprinklesClassName = gridSprinkles({
-    ...restProps,
-    gap,
-    gridTemplateColumns: columns ? columns : 'auto',
-  });
+  const utilityClassNames = getClassNames({ gap, columns, ...restProps });
 
-  const classNames = ['grid', sprinklesClassName, className]
+  const classNames = ['canon-grid', utilityClassNames, className]
     .filter(Boolean)
     .join(' ');
 
-  return createElement(
-    'div',
-    {
-      className: classNames,
-      style,
-    },
+  return createElement('div', {
+    ref,
+    className: classNames,
+    style,
     children,
-  );
-};
-
-const GridItem = (props: GridItemProps) => {
-  const { children, rowSpan, colSpan, start, end, className, style } = props;
-
-  const sprinklesClassName = gridItemSprinkles({
-    rowSpan,
-    colSpan,
-    start,
-    end,
   });
+});
 
-  const classNames = ['grid-item', sprinklesClassName, className]
+const GridItem = forwardRef<HTMLDivElement, GridItemProps>((props, ref) => {
+  const { children, className, style, ...restProps } = props;
+
+  const utilityClassNames = getClassNames(restProps);
+
+  const classNames = ['grid-item', utilityClassNames, className]
     .filter(Boolean)
     .join(' ');
 
-  return createElement('div', { className: classNames, style }, children);
-};
+  return createElement('div', {
+    ref,
+    className: classNames,
+    style,
+    children,
+  });
+});
 
-Grid.Item = GridItem;
+/** @public */
+export const Grid = Object.assign(GridBase, { Item: GridItem });