+++ packages/app-next-example-plugin/CHANGELOG.md
@@ -1,5 +1,21 @@
 # app-next-example-plugin
 
+## 0.0.18
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/frontend-plugin-api@0.9.3
+  - @backstage/core-components@0.16.2
+
+## 0.0.18-next.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/core-components@0.16.2-next.2
+  - @backstage/frontend-plugin-api@0.9.3-next.2
+
 ## 0.0.18-next.1
 
 ### Patch Changes