+++ plugins/auth-backend-module-guest-provider/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @backstage/plugin-auth-backend-module-guest-provider
 
+## 0.2.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/plugin-auth-node@0.5.5
+  - @backstage/backend-plugin-api@1.1.0
+  - @backstage/errors@1.2.6
+  - @backstage/catalog-model@1.7.2
+
+## 0.2.3-next.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/backend-plugin-api@1.1.0-next.2
+  - @backstage/errors@1.2.6-next.0
+  - @backstage/plugin-auth-node@0.5.5-next.2
+  - @backstage/catalog-model@1.7.2-next.0
+
 ## 0.2.3-next.1
 
 ### Patch Changes