+++ plugins/techdocs-react/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @backstage/plugin-techdocs-react
 
+## 1.2.12
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/core-components@0.16.2
+  - @backstage/catalog-model@1.7.2
+  - @backstage/config@1.3.1
+  - @backstage/core-plugin-api@1.10.2
+  - @backstage/version-bridge@1.0.10
+
+## 1.2.12-next.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/catalog-model@1.7.2-next.0
+  - @backstage/config@1.3.1-next.0
+  - @backstage/core-components@0.16.2-next.2
+  - @backstage/core-plugin-api@1.10.2-next.0
+  - @backstage/version-bridge@1.0.10
+
 ## 1.2.12-next.1
 
 ### Patch Changes